Move `view.ts` out (#141174)
