+++ src/vs/editor/browser/view.ts
